book_site/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
djoser
django-cors-headers
pillow
six